BLPDemand.jl

BLPDemand.delta

BLPDemand.demandmoments

BLPDemand.dsharedp

BLPDemand.eqprices

BLPDemand.loadblpdata

BLPDemand.share

BLPDemand.sharep

BLPDemand.simulateBLP

BLPDemand.simulateIVRClogit

BLPDemand.supplymoments

BLPDemand.delta — Method

function delta(s, x, ν, σ; tol=sqrt(eps(eltype(s))), maxiter=1000)

Solves for δ in s = share(δ, ...) using contraction mapping iteration.

See also: share

BLPDemand.demandmoments — Method

function demandmoments(β::AbstractVector,σ::AbstractVector,
                       s::AbstractMatrix, 
                       x, ν, ivdemand)

Demand side moments in BLP model.

Arguments

β length K (=number characteristics) vector of average tastes for characteristics

σ length K vector of standard deviations of tastes

s J × T matrix of market shares

x K × J × T array of product characteristics

ν K × S × T array of draws of ν

ivdemand L × J × T array of instruments. Identification requries L ≥ K.

Returns L × J array of moments.

See also: share, delta, simulateRCIVlogit

BLPDemand.dsharedp — Method

function dsharedp(β::AbstractVector,
              σ::AbstractVector,
              p::AbstractVector,
              x::AbstractMatrix,
              ν::AbstractMatrix,
              ξ::AbstractVector)

Compute market shares and their derivatives in BLP random coefficients demand model.

See sharep for argument details.

Returns

s vector J market shares

ds J × J Jacobian matrix with ds[l,j] = ∂s[l]/∂p[j]

Λ J × J diagonal matrix with ds = Λ - Γ

Γ J × J matrix

See eqprices for usage of Λ and Γ.

BLPDemand.eqprices — Method

function eqprices(mc::AbstractVector,
                  β::AbstractVector, σ::AbstractVector,
                  ξ::AbstractVector,
                  x, ν;
                  firmid= 1:length(mc),
                  tol=sqrt(eps(eltype(mc))),
                  maxiter=10000)

Compute equilibrium prices in BLP model using ζ contraction method of Morrow & Skerlos (2011).

Arguments

mc vector of J marginal costs

β vector of K taste coefficients

σ vector of K taste standard deviations

ξ vector of J demand shocks

x (K-1) × J exogenous product characteristics

ν K × S × T array of draws of ν

firmid= (1:J) .* fill(1, T)' identifier of firm producing each good. Default value assumes each good is produced by a different firm.

tol convergence tolerance

maxiter maximum number of iterations.

BLPDemand.share — Method

function share(δ::AbstractVector, σ::AbstractVector, x::AbstractMatrix, ν::AbstractMatrix)

Compute market shares in BLP random coefficients demand model.

Market shares are given by

\[ s_{j} = \int \frac{e^{\delta_j + x_j σ ν}}{1+\sum_{i=1}^J e^{\delta_i + x_i σ ν}} dF_ν(ν)\]

Arguments:

δ vector of length J= number of products

σ vector of length K= number of characteristics

x K × J matrix of product characteristics

ν K × S matrix of simulation draws for integration

Returns vector of length J market shares.

See also: delta

BLPDemand.sharep — Method

function sharep(β::AbstractVector, σ::AbstractVector, p::AbstractVector, x::AbstractMatrix, ν::AbstractMatrix, ξ::AbstractVector)

Compute market shares in BLP random coefficients demand model.

Market shares are given by

\[ s_{j} = \int \frac{e^{\delta_j + x_j σ ν}}{1+\sum_{i=1}^J e^{\delta_i + x_i σ ν}} dF_ν(ν)\]

where

\[\delta_j = β[1]*p[j] + x[:,j]' *β[2:end] + ξ[j]\]

Arguments:

β vector of length K= number of characteristics

σ vector of length K= number of characteristics

p J vector of prices

x (K-1) × J matrix of exogenous product characteristics

ν K × S matrix of simulation draws for integration

ξ J vector of demand shocks

Returns vector of length J market shares.

See also: share, dsharedp

BLPDemand.simulateBLP — Method

function simulateBLP(T, β, σ, γ, S)

Simulates a BLP demand and supply model.

Arguments

T::Integer number of markets

β::AbstractVector with length(β)=K, average tastes for characteristics. The first characteristic will be endogeneous (price)

σ::AbstractVector with length(σ)=K, standard deviation of tastes for characteristics

γ::AbstractVector marginal cost coefficients

S number of simulation draws to calculate market shares

Returns

x array of product characteristics with size(x) = (K, J, T)

z array of instruments with size(z) = (L, J, T)

s matrix of market shares size(s) = (J,T)

ν random draws used to compute market shares, size(ν) = (K,S,T)

ξ market demand shocks, size(ξ) = (J,T)

BLPDemand.simulateIVRClogit — Method

function simulateIVRClogit(T, β, σ, π, ρ, S)

Simulates a random coefficients logit model with endogeneity.

Arguments

T::Integer number of markets

β::AbstractVector with length(β)=K, average tastes for characteristics.

σ::AbstractVector with length(σ)=K, standard deviation of tastes for characteristics

π::AbstractMatrix with size(π) = (L, K, J), first stage coefficients

ρ::Number strength of endogeneity

S number of simulation draws to calculate market shares

Returns

x array of product characteristics with size(x) = (K, J, T)

z array of instruments with size(z) = (L, J, T)

s matrix of market shares size(s) = (J,T)

ν random draws used to compute market shares, size(ν) = (K,S,T)

ξ market demand shocks, size(ξ) = (J,T)

BLPDemand.supplymoments — Method

function supplymoments(γ::AbstractVector, β::AbstractVector, σ::AbstractVector,
                       ξ::AbstractVector, x, ν, w, ivsupply; 
                       firmid= (1:size(ξ,1)) .* fill(1, size(ξ,2))')

Supply side moments in a BLP model. Assumes that marginal cost is log linear,

\[c_{jt} = \exp(w_{jt}'γ + ω_{jt})\]

and prices are Bertrand-Nash

\[c_{t} = p_{t} + (∂s/∂p)^{-1} s\]

where ∂s/∂p is the Jacobian of shares with respect to prices, but with the j,l entry set to zero unless goods j and l are produced by the same firm.

Arguments

γ marginal cost coefficients

β length K (=number characteristics) vector of average tastes for characteristics

σ length K vector of standard deviations of tastes

ξ J × T matrix of market demand shocks

x K × J × T array of product characteristics

ν K × S × T array of draws of ν

w length(γ) × J × T array of cost shifters

ivsupply M × J × T array of instruments. Identification requries M ≥ length(γ).

firmid= (1:J) .* fill(1, T)' identifier of firm producing each good. Default value assumes each good is produced by a different firm.

Returns L × J array of moments.

BLPDemand.loadblpdata — Method

function loadblpdata()

Loads data from Berry, Levinsohn, and Pakes (1999).

Returns a DataFrame.